requirements.txt
webob
psutil
PasteDeploy
pyyaml
galaxy-job-metrics>=19.9.0.dev1
galaxy-objectstore>=19.9.0.dev1
galaxy-tool-util>=19.9.0.dev3
paramiko

## Uncomment if using DRMAA queue manager.
#drmaa

## To use alternative Python web servers.
#circus
#chaussette
# -or-
#git+git://github.com/mozilla-services/chaussette.git
